Fix a few signed/unsigned integer comparisons.

+++ regcomp.c
@@ -503,7 +503,7 @@ int starordinary;		/* is a leading * an ordinary character? */
 		i = (c&~BACKSL) - '0';
 		assert(i < NPAREN);
 		if (p->pend[i] != 0) {
-			assert(i <= p->g->nsub);
+			assert((size_t) i <= p->g->nsub);
 			EMIT(OBACK_, i);
 			assert(p->pbegin[i] != 0);
 			assert(OP(p->strip[p->pbegin[i]]) == OLPAREN);
@@ -1179,7 +1179,7 @@ freeset(p, cs)
 register struct parse *p;
 register cset *cs;
 {
-	register int i;
+	register size_t i;
 	register cset *top = &p->g->sets[p->g->ncsets];
 	register size_t css = (size_t)p->g->csetsize;
 
@@ -1205,7 +1205,7 @@ register struct parse *p;
 register cset *cs;
 {
 	register uch h = cs->hash;
-	register int i;
+	register size_t i;
 	register cset *top = &p->g->sets[p->g->ncsets];
 	register cset *cs2;
 	register size_t css = (size_t)p->g->csetsize;
@@ -1238,7 +1238,7 @@ firstch(p, cs)
 register struct parse *p;
 register cset *cs;
 {
-	register int i;
+	register size_t i;
 	register size_t css = (size_t)p->g->csetsize;
 
 	for (i = 0; i < css; i++)
@@ -1257,7 +1257,7 @@ nch(p, cs)
 register struct parse *p;
 register cset *cs;
 {
-	register int i;
+	register size_t i;
 	register size_t css = (size_t)p->g->csetsize;
 	register int n = 0;
